--- randow.ts
@@ -1,6 +1,0 @@
-export  class  Rando  {
-    
-    public  static  randomizar ( inicio : number ,  fim : number ) : number  {
-        return  inicio  +  Math.random()
-    }
-}
