--- saul.ts
@@ -1,13 +1,0 @@
-import { LutadorBase } from "./lutadores";
-export default class Saul extends LutadorBase {
-    protected media: number;
-
-    constructor(nome: string, categoria: string, defesa: number, forca: number, resistencia: number, velocidade: number, idade: number, media: number) {
-        super(nome, categoria, defesa, forca, resistencia, velocidade, idade, media);
-        this.media= media;
-
-    }
-    saul = this.media
-
-}
-
