+++ src/randow/randow.ts
@@ -1,0 +1,6 @@
+export  class  Rando  {
+    
+    public  static  randomizar ( inicio : number ,  fim : number ) : number  {
+        return  inicio  +  Math.random()
+    }
+}
